requirements.txt
asgiref==3.8.1
boto3==1.35.32
botocore==1.35.32
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.1.0
Django==4.2.16
django-cors-headers==4.4.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
firebase-admin==6.5.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
stripe==11.5.0
tweepy==4.14.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.6
whitenoise==6.6.0
django-markdownx==4.0.7
django-filter==24.1